Move PredictionsAndTanimotoScores to separate file
