+++ ms2deepscore/benchmarking/plot_heatmaps.py
@@ -1,8 +1,8 @@
 import numpy as np
 from matplotlib import pyplot as plt
 
-from ms2deepscore.benchmarking.CalculateScoresBetweenAllIonmodes import CalculateScoresBetweenAllIonmodes, \
-    PredictionsAndTanimotoScores
+from ms2deepscore.benchmarking.CalculateScoresBetweenAllIonmodes import CalculateScoresBetweenAllIonmodes
+from ms2deepscore.benchmarking.PredictionsAndTanimotoScores import PredictionsAndTanimotoScores
 
 
 def create_3_heatmaps(pairs: CalculateScoresBetweenAllIonmodes, nr_of_bins):
