+++ tests/test_plotting.py
@@ -1,10 +1,8 @@
-import random
 import pytest
-import numpy as np
 from matplotlib import pyplot as plt
 
-from ms2deepscore.benchmarking.CalculateScoresBetweenAllIonmodes import PredictionsAndTanimotoScores, \
-    CalculateScoresBetweenAllIonmodes
+from ms2deepscore.benchmarking.CalculateScoresBetweenAllIonmodes import CalculateScoresBetweenAllIonmodes
+from ms2deepscore.benchmarking.PredictionsAndTanimotoScores import PredictionsAndTanimotoScores
 from ms2deepscore.benchmarking.plot_average_per_bin import plot_average_per_bin
 from ms2deepscore.benchmarking.plot_heatmaps import create_3_heatmaps
 from ms2deepscore.benchmarking.plot_rmse_per_bin import (
